rename gitpush task to gitcommit
