+++ internal/task/gitcommit/commit_test.go
@@ -20,7 +20,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package gitpush
+package gitcommit
 
 import (
 	"io/ioutil"
